+++ R/vartymetric.R
@@ -314,7 +314,7 @@ print.mev_thselect_vmetric <-
 #' @param B number of simulations for variability of estimation
 #' @param probs quantile levels for intervals.
 #' @param ... additional arguments, currently ignored
-#' @return NULL; the function is used to produce a plot
+#' @rdname thselect.vmetric
 plot.mev_thselect_vmetric <-
   function(
     x,